+++ src/generator/index.ts
@@ -1,11 +1,7 @@
 import { OpenAPIV3 } from 'openapi-types';
 
 import { OpenApiRouter } from '../types';
-import {
-  builtInZodComponentGenerator,
-  zodComponentGenerator,
-  zodComponentRegistry,
-} from '../utils/registry';
+import { zodComponentProcessor } from '../utils/components';
 import { getOpenApiPathsObject } from './paths';
 import { errorResponseObject } from './schema';
 
@@ -46,13 +42,7 @@ export const generateOpenApiDocument = (
     paths: getOpenApiPathsObject(appRouter, Object.keys(securitySchemes)),
     components: {
       securitySchemes,
-      schemas: {
-        ...(zodComponentGenerator === 'built-in'
-          ? builtInZodComponentGenerator()
-          : zodComponentGenerator !== 'none' && zodComponentRegistry
-          ? zodComponentGenerator(zodComponentRegistry)
-          : {}),
-      },
+      schemas: zodComponentProcessor?.generateSchemas?.(),
       responses: {
         error: errorResponseObject,
       },
